pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi>=0.124.4",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=5.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "uvicorn>=0.38.0",
]
